Implementing redux to login
