+++ webapp/src/app/core/actions/login.action.js
@@ -1,0 +1,9 @@
+import types from '../types'
+
+export function login({ email, password }) {
+  return {
+    email,
+    password,
+    type: types.login.USER_LOGIN,
+  };
+}
