+++ webapp/src/app/core/reducers/index.js
@@ -1,7 +1,10 @@
 import { combineReducers } from 'redux'
 import { routerReducer } from "react-router-redux"
 
+import login from './login.reducer'
+
 const beagleboxapp = combineReducers({
+  login,
   routing: routerReducer,
 })
