+++ webapp/src/app/core/reducers/login.reducer.js
@@ -1,0 +1,31 @@
+import types from "../types"
+
+const initialState = {
+  user: {
+    name: '',
+    email: '',
+    password: '',
+    admin: false,
+  }
+};
+
+export default function login(state=initialState, action) {
+  switch (action.type) {
+    case types.login.USER_LOGIN: {
+      state = {
+        ...state,
+        user: {
+          ...state.user,
+          email: action.email,
+          password: action.password,
+        }
+      };
+
+      break;
+    }
+    default: {
+      break;
+    }
+  }
+  return state;
+};
