+++ webapp/src/app/core/types/index.js
@@ -1,9 +1,11 @@
+import login from './login.type'
+
 function combineTypes(typeArray) {
   return Object.assign({}, ...typeArray);
 }
 
 const types = combineTypes([
-
+  login,
 ]);
 
 export default types;
