+++ webapp/src/app/core/types/login.type.js
@@ -1,0 +1,7 @@
+const login = {
+  login: {
+    USER_LOGIN: "USER_LOGIN",
+  }
+};
+
+export default login;
