+++ webapp/src/app/view/homepage/index.js
@@ -1,35 +1,18 @@
-import React, { Component } from 'react'
-import { Grid, Row, Col } from 'react-flexbox-grid'
+import { connect } from "react-redux"
+import HomepageComponent from "./homepage"
 
-import {Card, CardHeader, CardText} from 'material-ui/Card'
+const stateToProps = (state) => {
+  return {
 
-import HomepageForm from './form'
-import assets from './assets'
-import './homepage.css'
-
-export default class Homepage extends Component {
-  render() {
-    return (
-      <Grid className="homepage-container" fluid>
-        <Row className="row-fluid homepage-content">
-          <Col className="col-fluid col-homepage" md={6} sm={12} xs={12}>
-            <img className="img-homepage" src={assets.beaglebox} alt={"BeagleBox logo"} />
-          </Col>
-          <Col className="col-fluid col-homepage" md={6} sm={12} xs={12}>
-            <Card className="homepage-card-container">
-              <CardHeader
-                className="homepage-card-title"
-                title="Seja bem-vindo"
-                showExpandableButton={false}
-                initiallyExpanded={true} />
-              <CardText
-                className="homepage-card-text" >
-                <HomepageForm />
-              </CardText>
-            </Card>
-          </Col>
-        </Row>
-      </Grid>
-    );
   }
-}
+};
+
+const dispatchToProps = (dispatch) => {
+  return {
+
+  }
+};
+
+const Homepage = connect(stateToProps, dispatchToProps)(HomepageComponent);
+
+export default Homepage;
